Made methods for QuantityItemDAO
Also converted all instances of "long" to "int" since IDs are stored in the database as "int" values due to limited space.

+++ src/main/java/com/mycompany/CodeforSefinal/InvoicePageController.java
@@ -85,7 +85,7 @@ public class InvoicePageController implements Initializable {
     @FXML
     private void handleAddItem(ReferenceItem rItem) {
         //Setup
-        Long itemId = rItem.getItemId();
+        int itemId = rItem.getItemId();
         String itemName = rItem.getName();
         Double itemPrice = rItem.getPrice();
         String itemPriceText = ""+rItem.getPrice();
